relinkPrivateKeys inside arrayContainerDefaults
